Move the email classes to its own namespace
