+++ src/EventManagement.Services/Messaging/IEmailSender.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace losol.EventManagement.Services
+namespace losol.EventManagement.Services.Messaging
 {
     public interface IEmailSender
     {
