+++ src/EventManagement.Services/Messaging/SendGridOptions.cs
@@ -1,0 +1,10 @@
+namespace losol.EventManagement.Services.Messaging
+{
+	public class SendGridOptions
+	{
+		public string User { get; set; }
+		public string Key { get; set; }
+		public string EmailAddress { get; set; }
+		public string Name { get; set; }
+	}
+}
