--- src/EventManagement.Services/SendGridOptions.cs
@@ -1,7 +1,0 @@
-public class SendGridOptions
-{  
-	public string User { get; set; }
-    public string Key { get; set; }
-	public string EmailAddress { get; set; }
-	public string Name { get; set; }
-}
