+++ src/EventManagement.Web/Pages/Account/ForgotPassword.cshtml.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using losol.EventManagement.Domain;
 using losol.EventManagement.Services;
+using losol.EventManagement.Services.Messaging;
 
 namespace losol.EventManagement.Pages.Account
 {
