+++ src/EventManagement.Web/Pages/Account/Register.cshtml.cs
@@ -10,6 +10,7 @@ using Microsoft.Extensions.Logging;
 using losol.EventManagement.Domain;
 using losol.EventManagement.Services;
 using losol.EventManagement.ViewModels;
+using losol.EventManagement.Services.Messaging;
 
 namespace losol.EventManagement.Pages.Account
 {
