+++ src/EventManagement.Web/Pages/Register/Event.cshtml.cs
@@ -7,6 +7,7 @@ using losol.EventManagement.Domain;
 using losol.EventManagement.Infrastructure;
 using losol.EventManagement.Pages.Account;
 using losol.EventManagement.Services;
+using losol.EventManagement.Services.Messaging;
 using losol.EventManagement.ViewModels;
 using Mapster;
 using Microsoft.AspNetCore.Hosting;
